refacto: reorganize StudentList folder
