--- src/components/StudentSection/StudentEventsList/StudentEventsList-styles.less
@@ -1,30 +1,0 @@
-.event-card {
-  width: 290px;
-  position: relative;
-  height: 200px;
-}
-
-.card-title {
-  position: absolute;
-  top: 15%;
-  left: 25%;
-  font-size: 25px;
-  color: #fff;
-} 
-
-.event-favorite {
-  position: absolute;
-  right: 5%;
-  bottom: 90%;
-  font-size: 25px;
-  color: #fff;
-}
-
-.card-date {
-  display: flex;
-  justify-content: right;
-}
-
-.img-cover {
-  filter: brightness(60%);
-}
