+++ src/components/StudentSection/StudentEventsList/_components/EventCard/EventCard-styles.less
@@ -1,0 +1,29 @@
+.event-card {
+  width: 290px;
+  position: relative;
+}
+
+.card-title {
+  position: absolute;
+  top: 15%;
+  left: 25%;
+  font-size: 25px;
+  color: #fff;
+} 
+
+.event-favorite {
+  position: absolute;
+  right: 5%;
+  bottom: 90%;
+  font-size: 25px;
+  color: #fff;
+}
+
+.card-date {
+  display: flex;
+  justify-content: right;
+}
+
+.img-cover {
+  filter: brightness(60%);
+}
