+++ src/components/StudentSection/StudentEventsList/_components/EventCard/EventCard-types.tsx
@@ -1,0 +1,6 @@
+import { IEventFormFields } from "../../../../Events/type"
+
+export interface IEventCardProps {
+  event: IEventFormFields
+  onClick: () => void
+}
